pyproject.toml
[tool.poetry]
name = "simple-benchmark"
version = "0.1.0"
description = ""
authors = ["Alberto Spina <alberto.spina.1996@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.7"
pydantic = "^2.4.2"
pyyaml = "^6.0.1"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
pysqlite3 = "^0.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
